pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "gif-to-sprite"
version = "0.2"
authors = [
  { name="githubcatw", email="17374742+githubcatw@users.noreply.github.com" },
]
description = "Converts a GIF to a sprite sheet."
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
  'Pillow',
]

[project.urls]
Homepage = "https://github.com/githubcatw/gif-to-sprite"
Issues = "https://github.com/githubcatw/gif-to-sprite/issues"
